refactor(#902): extract FixPack helper to reduce test boilerplate
